Fixed unbelievably stupid play.
evaluationsArray=[searchingIndex[0]]=[] in wrong place, meaning level 5
played like level 4, only having a search depth of 1 rather than a
nested array of evaluations. This is not a update as such and having
this as 0.7.1 would be a silly waste of space and time. 0.7.0 was
released in a rush, and 0.7.1 plans to be simply a neater version, not
what 0.7.0 was meant to be and now is.